### 2023-09-05 #5805 @editorialbot
[PRE REVIEW]: LIBSsa: an open source software for analyzing LIBS spectra
**Submitting author:** @kstenio (<a href="http://orcid.org/0000-0002-1407-9712">Kleydson Stenio</a>)
**Repository:** https://github.com/kstenio/libssa/
**Branch with paper.md** (empty if default branch): joss
**Version:** v2.1.0
