### 2023-09-05 da88e50
LIBSsa version: 2.1 [JOSS paper attempt #10]
paper.bib
* Added doi to some of the articles.
