+++ paper.bib
@@ -22,6 +22,7 @@
   pages={435--440},
   year={2011},
   publisher={Elsevier},
+  doi = {10.1016/j.talanta.2011.04.001},
   url = {https://doi.org/10.1016/j.talanta.2011.04.001},
 }
 
@@ -33,6 +34,7 @@
   pages={195--202},
   year={2016},
   publisher={Elsevier},
+  doi = {10.1016/j.geoderma.2015.09.018},
   url = {https://doi.org/10.1016/j.geoderma.2015.09.018},
 }
 
@@ -44,6 +46,7 @@
   pages={118--126},
   year={2018},
   publisher={Elsevier},
+  doi = {10.1016/j.microc.2018.05.008},
   url = {https://doi.org/10.1016/j.microc.2018.05.008},
 }
 
@@ -55,6 +58,7 @@
   pages={1886--1896},
   year={2018},
   publisher={Springer},
+  doi = {10.1007/s12161-017-1086-9},
   url = {https://doi.org/10.1007/s12161-017-1086-9},
 }
 
@@ -67,6 +71,7 @@
   pages={945--956},
   year={2018},
   publisher={Royal Society of Chemistry},
+  doi = {10.1039/C8JA00076J},
   url = {https://doi.org/10.1039/C8JA00076J},
 }
 
@@ -78,6 +83,7 @@
   pages={65--73},
   year={2018},
   publisher={Elsevier},
+  doi = {10.1016/j.trac.2018.08.003},
   url = {https://doi.org/10.1016/j.trac.2018.08.003},
 }
 
@@ -90,6 +96,7 @@
   pages={78--82},
   year={2016},
   publisher={Royal Society of Chemistry},
+  doi = {10.1039/C5AY01615K},
   url={https://doi.org/10.1039/C5AY01615K}
 }
 
@@ -102,6 +109,7 @@
   pages={2170--2176},
   year={2014},
   publisher={Optica Publishing Group},
+  doi = {10.1364/AO.53.002170},
   url={https://doi.org/10.1364/AO.53.002170}
 }
 
@@ -114,6 +122,7 @@
   pages={77--83},
   year={2021},
   publisher={Royal Society of Chemistry},
+  doi = {10.1039/D0AY02008G},
   url={https://doi.org/10.1039/D0AY02008G}
 }
 
@@ -126,6 +135,7 @@
   pages={4219--4229},
   year={2022},
   publisher={Royal Society of Chemistry},
+  doi = {10.1039/D2AY01424F},
   url={https://doi.org/10.1039/D2AY01424F}
 }
 
@@ -138,6 +148,7 @@
   pages={2005--2014},
   year={2016},
   publisher={Royal Society of Chemistry},
+  doi = {10.1039/C6JA00224B},
   url={https://doi.org/10.1039/C6JA00224B}
 }
 
@@ -150,6 +161,7 @@
   pages={1246--1253},
   year={2022},
   publisher={Royal Society of Chemistry},
+  doi = {10.1039/D1AY02101J},
   url={https://doi.org/10.1039/D1AY02101J}
 }
 
@@ -162,6 +174,7 @@
   pages={1552--1565},
   year={2004},
   publisher={IEEE},
+  doi = {10.1109/TGRS.2004.830549},
   url={https://doi.org/10.1109/TGRS.2004.830549}
 }
 
@@ -174,6 +187,7 @@
   pages={764--766},
   year={2013},
   publisher={Elsevier},
+  doi = {10.1016/j.jesp.2013.03.013},
   url = {https://doi.org/10.1016/j.jesp.2013.03.013},
 }
 
@@ -186,6 +200,7 @@
   pages={357--362},
   year={2020},
[…]
