### 2023-09-06 #5805 @kstenio
Hello @Kevin-Mattheus-Moerman, thank you very much for the support in this review. I've already added the missing DOIs in the paper.bib file. Regarding potential reviewers, I've checked the [recommended list](https://reviewers.joss.theoj.org/lookup) and found some of them with repositories (and/or forks) with subject aligned to mine (spectroscopy): tillsten, ivergara and ksunden.
