### 2023-09-28 #5805 @csoneson
Ping @Tillsten, @ivergara, @ksunden - would you be willing to review this submission for JOSS?

> 👋🏻 @Tillsten, @ivergara, @ksunden - would you be interested in reviewing this submission for JOSS?
>
> > LIBSsa: an open source software for analyzing LIBS spectra
> > #5805
>
> Thanks in advance!
