### 2023-10-07 #5805 @csoneson
👋🏻 @jwuttke, @tpeulen, @alberto-battistel - would you be interested in reviewing this submission for JOSS?

> LIBSsa: an open source software for analyzing LIBS spectra
> https://github.com/openjournals/joss-reviews/issues/5805

Thanks in advance!
