### 2023-10-15 #5805 @csoneson
👋🏻 @adrien-matta, @domna, @jwuttke, @tpeulen, @alberto-battistel - would you be interested in reviewing this submission for JOSS?

LIBSsa: an open source software for analyzing LIBS spectra
https://github.com/openjournals/joss-reviews/issues/5805

JOSS is a free, open-source, community-driven and developer-friendly online journal. The review process is unique: it takes place in a GitHub issue, is open, and author-reviewer-editor conversations are encouraged.

JOSS reviews involve downloading and installing the software, and inspecting the repository and submitted paper for key elements. See https://joss.readthedocs.io/en/latest/review_criteria.html.

Thanks in advance!
