### 2023-10-18 #5805 @Tillsten
Is there example data given? If so, I can try do a review
