### 2023-10-18 #5805 @kstenio
Hi @Tillsten, thank you very much for your availability! Yes, I have some spectra/example data I can provide in order to test software usability. If it is OK for @csoneson, I can create a short lived branch (based on current "joss" branch) with compressed files.
