### 2023-10-18 #5805 @csoneson
@kstenio - sure, that should work (unless the files are too big for GitHub). In general though, it would be good to have some small example data provided with the package, or somewhere else easily accessible, so that a new user could quickly test e.g. that their installation works and see what the output files would look like.
