### 2023-10-18 #5805 @kstenio
Great @csoneson. I've created a new branch, [joss_sample_data](https://github.com/kstenio/libssa/tree/joss_sample_data), and inside the folder [data](https://github.com/kstenio/libssa/tree/joss_sample_data/data) you will able to find the compressed (**7z**) sample spectra and some instructions.

Regarding the small data, I see no problems in including a minor sample set in the official repository.
