### 2023-10-19 #5961 @editorialbot
[REVIEW]: LIBSsa: an open source software for analyzing LIBS spectra
**Submitting author:** @kstenio (<a href="http://orcid.org/0000-0002-1407-9712">Kleydson Stenio</a>)
**Repository:** https://github.com/kstenio/libssa/
**Branch with paper.md** (empty if default branch): joss
**Version:** v2.2.0
