### 2023-10-19 #5805 @adrien-matta
Dear Charlotte,

Thanks a lot for this opportunity. Unfortunately I am not available at this time to provide a review.

Best regards,
Adrien.

_________________________
• Dr Adrien Matta
• Chargé de Recherche CNRS
• [ ***@***.*** | ***@***.*** ] • +33 2 31 45 24 43
• LPC Caen, ENSICAEN, 6 Boulevard du Maréchal Juin, 14050, CAEN Cedex

De: "Charlotte Soneson" ***@***.***>
À: "openjournals/joss-reviews" ***@***.***>
Cc: "adrien matta" ***@***.***>, "Mention" ***@***.***>
Envoyé: Dimanche 15 Octobre 2023 11:57:55
Objet: Re: [openjournals/joss-reviews] [PRE REVIEW]: LIBSsa: an open source software for analyzing LIBS spectra (Issue #5805)

👋🏻 [ https://github.com/adrien-matta | @adrien-matta ] , [ https://github.com/domna | @domna ] , [ https://github.com/jwuttke | @jwuttke ] , [ https://github.com/tpeulen | @tpeulen ] , [ https://github.com/alberto-battistel | @alberto-battistel ] - would you be interested in reviewing this submission for JOSS?

LIBSsa: an open source software for analyzing LIBS spectra
[ https://github.com/openjournals/joss-reviews/issues/5805 | #5805 ]

JOSS is a free, open-source, community-driven and developer-friendly online journal. The review process is unique: it takes place in a GitHub issue, is open, and author-reviewer-editor conversations are encouraged.

JOSS reviews involve downloading and installing the software, and inspecting the repository and submitted paper for key elements. See [ https://joss.readthedocs.io/en/latest/review_criteria.html | https://joss.readthedocs.io/en/latest/review_criteria.html ] .

Thanks in advance!
