### 2023-10-20 1943fed
Update README.md
Revised Copyright section in README.md file
