### 2023-10-24 kstenio/libssa#2 @domna
Provide as python package
I think providing this program as a python package (maybe even via pypi if desired) would be a huge benefit for users, because they simply could do a `pip install libssa`.

I recommend to use the latest packaging methods from python which simply needs a [`pyproject.toml`](https://setuptools.pypa.io/en/latest/userguide/pyproject_config.html) file added to the repo. There is also a way to add [gui scripts](https://setuptools.pypa.io/en/latest/userguide/entry_point.html#gui-scripts) so you get an executable for your program (e.g., you could just call libssa or libssa.exe to execute the program) It can then also automatically build and shipped to pypi via the github actions, see [this](https://packaging.python.org/en/latest/guides/publishing-package-distribution-releases-using-github-actions-ci-cd-workflows/) for a reference.

Also a standard python [`.gitignore`](https://github.com/github/gitignore/blob/main/Python.gitignore) should be added.

It's also a good practice to list the dependencies unpinned in the pyproject.toml and let the requirements file being generated by [`pip-compile`](https://github.com/jazzband/pip-tools).
