### 2023-10-24 kstenio/libssa#3 @domna
README should be cleaner
I think the README could be made cleaner. I would not explain how to install python, but rather refer to a general tutorial. I would rather add some more ideas about your software, maybe a small example with a screenshot so people quickly understand what they are looking at and can quickly judge whether this will be useful to them.

I agree with #1 that the section about copyright and licensing should be clearer. It would also be useful to have a paragraph explaining how to contribute or how/where to contact if someone has an issue.

Here are some more comments on the different README sections:

### About the software
I would remove the header and place this directly under your heading (`LIBSsa2...`) so people directly start reading what the software is about.

### Features
I find this quite overwhelming here and would rather expect a small example and/or screenshot of a procedure of a software. Often people use gifs to showcase a simple use case with the software which I find really neat.

### Installing python + External libraries
As already stated I would remove this in favor of an example showcasing here. It's often not useful to have this and might be outdated at some point, so a reference to the official install instructions is the better choice. The libraries also don't matter to the inexperienced users and those more experienced typically know that they simply will find it in pyproject.toml or requirements.txt
