### 2023-10-24 #5961 @domna
The paper mentions Debora Milori as one of the authors of the software. However, there is no contribution or mention anywhere in the software or documentation. Could you @kstenio clarify the relationship and maybe add a paragraph in the acknowledgements that the research was carried out in the working group of Debora Milori (as I assume this was the case).

@csoneson What are the general JOSS guidelines here? It is stated that the people mentioned on the paper should've contributed to the software directly. However, in research we generally have PIs mentioned in the papers as well who typically don't contribute directly. Is this a contradiction to the JOSS rules and is there a way to express an associated person without a major contribution to the actual software?
