### 2023-10-24 kstenio/libssa#4 @domna
No (automated) tests
The software should contain at least basic (ideally automated with gh actions) tests to verify that it is working as intended.
The most used testing framework for such cases is probably pytest with [pytest-qt](https://pypi.org/project/pytest-qt/).
