### 2023-10-24 #5961 @csoneson
Hi @domna - please see [here](https://joss.readthedocs.io/en/latest/submitting.html#authorship) for the JOSS guidelines on authorship.
