### 2023-10-24 5e4518d
Update README.md
Addressed proper citing
