+++ README.md
@@ -117,8 +117,9 @@ the development of **LIBSsa 2** started.
 
 ## Acknowledgements
 
-The present software (up to version 2.0.99) was supported by the Coordination for the Improvement
-of Higher Education Personnel - Brazil (CAPES) - Finance Code 001, and by Embrapa Instrumentation.
+The present software (up to version 2.0.99) was supported by the Coordination for the Improvement 
+of Higher Education Personnel - Brazil (CAPES) - Finance Code 001, and by the Laboratory of Optics and Photonics 
+at Embrapa Instrumentation, under the supervision of Dr. Debora Milori.
 
 ## Warranty
 
