### 2023-10-24 #5961 @kstenio
> The paper mentions Debora Milori as one of the authors of the software. However, there is no contribution or mention anywhere in the software or documentation. Could you @kstenio clarify the relationship and maybe add a paragraph in the acknowledgements that the research was carried out in the working group of Debora Milori (as I assume this was the case).

Hi @domna, thank you for the highlight. In fact, Debora was my supervisor during my doctorate and PI/leader of working group (Laboratory of Optics and Photonics at Embrapa Instrumentation). I do mention Embrapa Instrumentation in the documentation, but did not mention Debora. To address this, I did a [small commit](https://github.com/kstenio/libssa/commit/5e4518d833449f9811cfd824638ccdba88e5dab2) in the **_joss_** branch citing her.

To clarify, she did not program during LIBSsa development, but placed an important role in it. Many core functionalities were added due to Debora's insights. Also, regarding the paper itself, she did help in writing and revising.

Finally, the **Author contributions** section in paper proof was added following [CRediT](https://beta.elsevier.com/researcher/author/policies-and-guidelines/credit-author-statement) guidelines (Contributor Roles Taxonomy).
