### 2023-10-24 #5961 @domna
> Hi @domna, thank you for the highlight. In fact, Debora was my supervisor during my doctorate and PI/leader of working group (Laboratory of Optics and Photonics at Embrapa Instrumentation). I do mention Embrapa Instrumentation in the documentation, but did not mention Debora. To address this, I did a [small commit](https://github.com/kstenio/libssa/commit/5e4518d833449f9811cfd824638ccdba88e5dab2) in the **_joss_** branch citing her.
>
> To clarify, she did not program during LIBSsa development, but placed an important role in it. Many core functionalities were added due to Debora's insights. Also, regarding the paper itself, she did help in writing and revising.

Ok, I think this is what is exactly coverd in in the JOSS guidelines as `[...]but active project direction and other forms of non-code contributions are`. So I think this is totally fine. Thank you for the explanation and update to the repo!
