### 2023-10-24 kstenio/libssa#2 @kstenio
Thank you for the feedback. Yes, I also think the software would have great benefit from being distributed as a Python package. It would be nice just using `pip install libssa` with all dependencies being taken care by pip. Unfortunately, my knowledge in this subject is very limited. I'll read instructions/links provided and see if I'm able to do it myself.

Either way, and I truly believe that cloning the repository (or downloading the zip), and running `pip install -r requirements.txt` is sufficiently simple for most users of the software (at least those familiar with pip).

Finally, the **.gitignore** file was added as requested in [this commit](https://github.com/kstenio/libssa/commit/79d8b9807ee244091a490a5340962ebb6bf75edd).
