### 2023-10-24 kstenio/libssa#4 @kstenio
Thank you again for the comments and feedback. I have to admit that testing is my Achilles' heel as I'm not very familiar with it. I've started the development of LIBSsa back in 2017, short time after I transitioned from MATLAB, and unit and integration tests wasn't so important as they are now. I only started studying TDD and the pytest library recently, just two months ago.

Nevertheless, I did extensive end to end tests during the whole process of developing the software. They simply weren't automated. Also, since my laboratory colleagues were using LIBSsa too, their feedback was used to do some fine tuning in functions/methods. There are plenty of conditionals and try/except statements to ensure inputs are correct before running a function, and that outputs are as they are supposed to be. E.g.: if you try to open a random txt file as a spectrum, you will receive a message box stating that the file structure is illegal.

Since tests were not a core focus, it is difficult for me, alone, to implement it. In this case, it would be great to have a contributor to the project focusing on this. While working on contributing guidelines (stated in issue #3), I think I'll add some sort of "features in need" section, and mention some areas where I know help would be appreciated.

Thanks for pointing out the pytest-qt library. I'll check its documentation.
