### 2023-10-24 kstenio/libssa#3 @kstenio
Thank you for the feedback. I'll reformulate the README.md file.

> It's often not useful to have this and might be outdated at some point...

In fact, I also think it is outdated, as I wrote many years ago and just updated as needed. But I did not knew how could be best practices. Having some external input is nice, and I'll think on how to simply the section.

> maybe a small example with a screenshot so people quickly understand what they are looking at and can quickly judge whether this will be useful to them

I like the idea of using pictures. In fact, the ones in the article proof does ring a bell. I'll think better on how to integrate them on the README.

> It would also be useful to have a paragraph explaining how to contribute or how/where to contact if someone has an issue.

I'm studying the creation of a [CONTRIBUTING](https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors) file to better address this concern.
