### 2023-10-25 df53b8b
Added CONTRIBUTING.md file with guidelines for community.
