+++ README.md
@@ -136,9 +136,7 @@ proper credit is given to the developer and/or this repository is mentioned. The
 To ensure the open-source nature of LIBSsa 2 and prevent any attempt to transform it into proprietary software, the source code
 of a previous version (2.0.99) has been registered with the Brazilian National Institute of Industrial Property (INPI-BR). 
 This registration provides an additional layer of protection in accordance with the principles of the 
-*[TRIPS Agreement]([url](https://www.wto.org/english/tratop_e/trips_e/intel2_e.htm))* and is documented under the process number *BR 51 2023 002165-2*.
-
-By taking these measures, we are committed to maintaining LIBSsa 2 as an open-source tool for the LIBS community.
+*[TRIPS Agreement](https://www.wto.org/english/tratop_e/trips_e/intel2_e.htm)* and is documented under the process number *BR 51 2023 002165-2*.
 
 ---
 
