### 2023-10-25 kstenio/libssa#3 @kstenio
I've already added proper guidelines for contributing. You may check the [CONTRIBUTING.md](https://github.com/kstenio/libssa/blob/joss/CONTRIBUTING.md) file. Tomorrow I'll try to update the **README.md** file as well (to be noticed, my timezone is GMT-3)
