### 2023-10-25 kstenio/libssa#2 @domna
If you like I can add a PR with basic packaging as a template to get you started. You only need to add your dependencies and metadata and it should be good to go.
