### 2023-10-25 kstenio/libssa#4 @domna
I can assist you in setting up some basic tests. I trust that the software is rigorously manually tested but please be aware that tests are not only there to show that the software works as intended from past behaviour but also ensures that the features don't break as soon as you make changes to the code (a typical workflow is: you find a bug, you write a test which fails on this bug, you fix it, test stays to ensure this doesn't happen again).
I understand that it is a lot of work to write tests so I won't request a huge test coverage for the JOSS review but I think there should be at least a test which checks that the gui works, data loads and maybe one of the major analysis steps works. I like to idea of adding a feature table people contributing to this repo can refer to. Often projects also use some labelling in their issues (like `Good first issue` or `important feature` or so). Do you think it is feasible for you to add these basic tests?
