### 2023-10-25 kstenio/libssa#3 @domna
> Thank you for the feedback. I'll reformulate the README.md file.
>
> > It's often not useful to have this and might be outdated at some point...
>
> In fact, I also think it is outdated, as I wrote many years ago and just updated as needed. But I did not knew how could be best practices. Having some external input is nice, and I'll think on how to simplify the section.

I think a good practice is to just link to the general python docs on how to install things or a good tutorial somewhere. It would also be nice if you state there which python version or software works with or at least on which it was tested (also a case for ci/cd testing - you can easily do a matrix test to check different versions).
>
> > maybe a small example with a screenshot so people quickly understand what they are looking at and can quickly judge whether this will be useful to them
>
> I like the idea of using pictures. In fact, the ones in the article proof does ring a bell. I'll think better on how to integrate them on the README.

Yes, right there are already these pictures. I think it would be nice to have one of those so I directly see how this software looks like.
>
> > It would also be useful to have a paragraph explaining how to contribute or how/where to contact if someone has an issue.
>
> I'm studying the creation of a [CONTRIBUTING](https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors) file to better address this concern.

I think this is a good way to go. Often projects also just have a small paragraph on the bottom of the readme where they explain how to contribute. This might be enough for you for a start. But if you like to write a whole contributing document even better :)
