### 2023-10-25 kstenio/libssa#3 @domna
> ... (to be noticed, my timezone is GMT-3)

No worries. I'm aware of the time difference and expect communication to be asynchronously.
