### 2023-10-25 kstenio/libssa#2 @kstenio
> If you like I can add a PR with basic packaging as a template to get you started. You only need to add your dependencies and metadata and it should be good to go.

This would be nice! Recently I've created my first [Python package on PyPI](https://pypi.org/project/psycopg-dbconn-class/) (it is a very simple one, but I did mostly to understand the packaging flow). However, I've used the **setup.py** approach, which I know it is being deprecated. Also, there are no tests there...

Having a template to follow, here on GitHub or even in the TestPyPI would be highly appreciated.
