+++ README.md
@@ -1,109 +1,89 @@
+<p align="center">
+	<img alt="libssa_logo" src="./pic/libssa.svg" width="500em">
+</p>
+
 # LIBSsa2 - Laser Induced Breakdown Spectroscopy spectra analyzer
 
 ### Current stable version: _2.1_
 
-## About the software
-
 **LIBSsa 2** is free software, licensed under the **GNU Affero General Public License version 3**, 
-and written in **Python 3.x** for analyzing **LIBS spectra**. It can read multiple formats of data inside raw/text files,
-import the spectra, and perform multiple analyses on the data. The environment (spectra, peaks, and results) may be saved into an encapsulated **lb2e** file.
+and written in **Python 3.x** for analyzing **LIBS spectra**. It can read several formats of data inside raw/text files,
+import the spectra, and perform multiple analyses on the data. The environment (spectra, peaks, and results) may also be saved 
+into an encapsulated **lb2e** file.
+
+<p align="center">
+	<img alt="libssa_mainwindow" src="./pic/libssa.png" width="900em"><br>
+	<i>LIBSsa main window</i>
+</p>
+
+For a full list of features implemented in LIBSsa, you may check the [FEATURES.md](./FEATURES.md) file but, in short, you can: 
+remove of outliers using SAM or MAD algorithms, perform full spectrum normalization (FSN), create correlation spectrum, 
+do peak isolation, peak fitting (Gaussian, Lorentzian, and Voigt), univariate linear models, multivariate partial least 
+squares regression (PLSR), principal components analysis (PCA), and obtain plasma temperature and electron density values 
+using Saha-Boltzmann (modified to LIBS) equation.
+
+Bellow are presented some of the graphics generated by LIBSsa.
+<p align="center">
+	<img alt="libssa_graphics" src="./pic/montage.png" width="900em"><br>
+	<i>Some treatments available in LIBSsa: peak isolation (a), peak fitting (b), PCA (c), and Saha-Boltzmann plot (d)</i>
+</p>
+
 Finally, users can also export the results of the analysis in many common formats (**txt**, **csv**, and **xlsx**).
 
-## Features
+## Setting up the environment
 
-1. **Importing Spectra**: Users can import data into LIBSsa 2 from any kind of raw/text data, including **txt**, **csv**, **esf**, and **ols**.
-There are also many options for how data can be loaded:
-   1. **Load methods**: How files are read by the program
-      1. **Multi**: One folder per sample, several files per sample
-      2. **Single**: One file per sample, multiple columns in each file
-   2. **Delimiters**: TAB ("**\t**"), SPACE ("**\s+**"), comma ("**,**"), and semicolon ("**;**")
-   3. **Header**: Number of rows to skip
-   4. **Wavelength and Counts**: Columns containing the wavelength and counts data
-   5. **Decimals**: If the user wants to round the loaded signal (for improved performance)
-   6. **FSN** (Full Spectrum Normalization): Can use signal information during the loading for normalization
-2. **Outliers Removal**: There are 2 algorithms for outliers removal in LIBSsa 2:
-   1. **SAM** (Spectral Angle Mapper): Removes outliers based on how different a sample is from the average
-   2. **MAD** (Median Absolute Deviation): Removes outliers based on a per-wavelength criterion
-3. **Correlation Spectrum**: If a reference is provided, calculates Pearson-R as a function of wavelength
-4. **Peak Isolation**: Can isolate multiple peaks at once
-5. **Peak Fitting**: Can perform peak fitting to obtain areas and heights for multiple functions/curves:
-   1. Lorentzian
[…]
