### 2023-10-26 kstenio/libssa#3 @kstenio
Hi, I updated the [README.md](https://github.com/kstenio/libssa/blob/joss/README.md) file!

Now I feel the information about the software is more intuitive thanks to the added pictures. I also moved the features section to a new file, [FEATURES.md](https://github.com/kstenio/libssa/blob/joss/FEATURES.md), and there is a tiny **Contributing** section pointing out to the [CONTRIBUTING.md](https://github.com/kstenio/libssa/blob/joss/CONTRIBUTING.md) file.

I also created a new section, [Setting up the environment](https://github.com/kstenio/libssa/tree/joss#setting-up-the-environment), which contains clear instructions on how obtain and install Python and libraries. I did not remove the mention to libraries because I like the idea on citing them to wider audiences.

I hope the changes are sufficient to close this issue.
