### 2023-10-26 kstenio/libssa#2 @kstenio
Hi, I have to say, after updating the **README.md** to address issue #3 (especially the creation of the new [Setting up the environment](https://github.com/kstenio/libssa/tree/joss#setting-up-the-environment) section), I have mixed opinions on this issue.

My point is, it would be nice to distribute LIBSsa on PyPI. It would be simpler for users to enter `pip install libssa` and run `./libssa` or `libssa.exe`. However, as a package, the source code of LIBSsa would be much more enclosed than it is now.

My primary aim for this project is to foster the community of LIBS users (and developers) being closely connected to the source code. By keeping the installation process as manual as it is now (cloning the repository, installing libraries with `pip install -r requirements.txt`, and running the source file with `python libssa.py`), I hope to encourage users to explore the code, gain a deeper understanding of how LIBSsa functions, and perhaps influence users to contribute to its development.

That said, I appreciate the convenience that packaging can offer. Therefore, I'm considering a compromise. It's a trade-off between convenience and keeping the source code easily accessible. Both approaches have their merits...
