### 2023-10-26 kstenio/libssa#3 @Tillsten
Hey, I am fully agree here @domna. Currently, it is not clear at what kind of user the instructions are aiming for. The Python installation instructions are aimed at beginners, but the rest is to terse for them. E.g. it is not mentioned how to download the package itself, that you have to be in the correct directory or how to start the software. I also think setting up some minimal packaging  would simplify the installation, even if you do not want the package to be on PyPI. Furthermore, especially beginners prefer a GUI like libssa, therefore making the installation a simple as possible is advantageous.
Btw, one way were the current instructions do fail is when a user just uses the latest Python, since a binary numpy wheel is still not available for 3.12. I like that you already mention are working python version, but I think the tested version should be a little bit more recent (3.11 seems to be working).
