### 2023-10-27 7ca430a
Addressing issues #1 and #3
README.md
* Added steps regarding cloning the repository and entering source folder
* Included references to other projects in Copyright section

CONTRIBUTING.md
* Added features in need for working with newer python versions and packaging
