### 2023-10-27 kstenio/libssa#3 @kstenio
Hi, thank you for the feedback. Addressing the issues:

> The Python installation instructions are aimed at beginners, but the rest is too terse for them.

In one of my latest commits, I removed the extensive material on installing Python on Linux (and Windows) in favor of a new, simpler [Setting up the environment](https://github.com/kstenio/libssa/tree/joss#setting-up-the-environment) section. Now, I only use one paragraph to clarify that subject, opting to reference the official instructions on how to install Python on Linux, Windows, and macOS, as suggested previously.

> E.g. it is not mentioned how to download the package itself, that you have to be in the correct directory or how to start the software.

I updated the [README.md](https://github.com/kstenio/libssa/blob/joss/README.md), adding instructions regarding the download of the **software zip** from GitHub or cloning the repository using `git`. Also, I've included a reference to `cd` into the software source folder _before_ running `pip` to install libraries (although this was already mentioned in the text).

> Furthermore, especially beginners prefer a GUI like libssa, therefore making the installation as simple as possible is advantageous.

I agree that packaging would have its benefits, but I think this shouldn't be the focus now. I'm much more inclined to invest efforts in implementing proper testing algorithms, as pointed out in issue #4. Regardless, I've added packaging as a possible improvement in the [CONTRIBUTING.md](https://github.com/kstenio/libssa/blob/joss/CONTRIBUTING.md) file. Either way, I believe the readme/instructions are much clearer and better now than they were _before_ the review process. In short, a user only has to run three commands to set up the environment: `git clone`, `cd`, and `pip install`.

> I like that you already mention a working python version, but I think the tested version should be a bit more recent (3.11 seems to be working).

It's always nice to have cutting-edge software running. However, my working environment is LMDE 5 (Linux Mint Debian Edition), which is based on **Debian 11**, a _super stable_ Linux distribution. LMDE 5 came with Python 3.9, and that's why this is the tested version. Debian 12 just arrived, and LMDE 6 was released a month ago. Another point is that Python 3.9 isn't dead yet: its lifespan is scheduled until [October 2025](https://peps.python.org/pep-0596/#lifespan), and the latest version, 3.9.18 (a security fix), was released just a [few months ago](https://www.python.org/downloads/release/python-3918/). This update (to 3.11) will require time, which I, alone, don't have. On the other hand, the software **is working** with the recommended settings, as well described in the readme/instructions. Nonetheless, to address the record of this subject, I've added this as a possible improvement too in the **CONTRIBUTING.md** file.

> Also, linking the example files for new users would be great.

I'll add the sample spectra (synthetic samples) from the sample_data branch to the main branch as well. :-)
