### 2023-11-02 kstenio/libssa#2 @domna
> Hi, I have to say, after updating the **README.md** to address issue #3 (especially the creation of the new [Setting up the environment](https://github.com/kstenio/libssa/tree/joss#setting-up-the-environment) section), I have mixed opinions on this issue.
>
> My point is, it would be nice to distribute LIBSsa on PyPI. It would be simpler for users to enter `pip install libssa` and run `./libssa` or `libssa.exe`. However, as a package, the source code of LIBSsa would be much more enclosed than it is now.
>

I understand your concern that you loose the direct access to the repository, but I believe that packaging lowers the entry barrier (many researchers are not used to version control systems!) for your package and a person wanting to contribute should be able to figure it out (for my projects I normally add a development install and "normal" install section).
Also PyPi is designed for open source software and is able to link to the git repository, issue tracker, contributing etc. Additionally, you can always install it from the git repo via `pip install git+your_repo_url` and pypi includes also an sdist which is a complete source distribution containing all your source files.

When I'm deciding whether I want to contribute to any open source project, available packaging is always a plus and I think many devs think similar about this.

> My primary aim for this project is to foster the community of LIBS users (and developers) being closely connected to the source code. By keeping the installation process as manual as it is now (cloning the repository, installing libraries with `pip install -r requirements.txt`, and running the source file with `python libssa.py`), I hope to encourage users to explore the code, gain a deeper understanding of how LIBSsa functions, and perhaps influence users to contribute to its development.
>
> That said, I appreciate the convenience that packaging can offer. Therefore, I'm considering a compromise. It's a trade-off between convenience and keeping the source code easily accessible. Both approaches have their merits...

I think you can still support both approaches: Create a package, upload it to pypi but have a broad section about setting it up for development from git. I think a good documentation and invitation to contribute + open list of issues is a much greater incentive for contributing to your project than forcing the users to downloading your package from git directly.
