+++ libssa/libssa2.py
@@ -1,0 +1,1646 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+#
+# Copyright (c) 2023 Kleydson Stenio (kleydson.stenio@gmail.com).
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU Affero General Public License as published
+# by the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU Affero General Public License for more details.
+#
+# You should have received a copy of the GNU Affero General Public License along
+# with this program.  If not, see <https://www.gnu.org/licenses/agpl-3.0.html>.
+
+
+# Imports
+import sys
+
+try:
+    import lzma
+    import pickle
+    import libssa.env.export as export
+    from time import time
+    from os import listdir
+    from pathlib import Path
+    from pandas import DataFrame
+    from markdown import markdown
+    from datetime import datetime
+    from traceback import print_exc
+    from psutil import virtual_memory
+    from libssa.env.spectra import Spectra, Worker
+    from libssa.env.gui.libssagui import LIBSsaGUI, changestatus
+    from libssa.env.imports import load, outliers, refcorrel, domulticorrel
+    from libssa.env.functions import (
+        isopeaks,
+        fitpeaks,
+        linear_model,
+        zeros,
+        column_stack,
+        pca_do,
+        pca_scan,
+        pls_do,
+        tne_do,
+        array,
+    )
+    from PySide6.QtWidgets import (
+        QApplication,
+        QMessageBox,
+        QMainWindow,
+        QTableWidgetItem,
+    )
+    from PySide6.QtCore import QThreadPool, QObject, QCoreApplication, Qt
+except (ImportError, ImportWarning) as err:
+    err_msg = str(err)
+    print(
+        f"\nYou have missing libraries to install.\n\n"
+        f"\tError message: {err_msg}\n\n"
+        f"Check the README.md for extra info.\n"
+    )
+    if "opengl" in err_msg.lower():
+        print(
+            'If you are under Linux Mint 20+ (or Ubuntu 20.04+), try running: "apt install libopengl0"'
+        )
+    elif "xcb" in err_msg.lower():
+        print(
+            'If you are under Linux Mint 20+ (or Ubuntu 20.04+), try running: "apt install libxcb-xinerama0 libxcb-cursor0"'
+        )
+    print_exc()
+    sys.exit(1)
+
+
+# LIBSsa main class
+class LIBSSA2(QObject):
+    """
+    This is LIBSsa main APP.
+
+    In it, we have all needed functions, actions and connects for the app to work properly.
+    """
+
+    def __init__(self, ui_file: str, logo_file: str):
+        # Checks if ui file exists and warn users if not
+        try:
+            super(LIBSSA2, self).__init__()
+            self.gui = LIBSsaGUI(ui_file, logo_file)
+        except Exception as ex:
+            QMessageBox.critical(
+                QMainWindow(None),
+                "Critical error!",
+                f"Can not initialize LIBSsa!"
+                f'<p>Error message: <b style="color: red">{str(ex)}</b></p>',
+            )
+            sys.exit(1)
+        else:
+            # If all is fine with ui, then starts to read other modules
+            self.spec = Spectra()
+            # Defines global variables
+            self.threadpool = QThreadPool()
+            self.parent = Path()
+            self.mbox = QMessageBox()
+            self.mode, self.delimiter = "", ""
+            self.cores, self.timer = 0, 0
[…]
