+++ MANIFEST.in
@@ -1,0 +1,2 @@
+recursive-include libssa/pic *
+recursive-include libssa *.ui
