### 2023-11-02 ff3e4f9
Adds basic qt test
