### 2023-11-08 c10166b
Update publish-as-package.yaml
Added/updated for TestPyPI
