+++ .github/workflows/publish-as-package.yaml
@@ -8,7 +8,7 @@
 
 name: Upload Python Package [Test]
 
-on: [pull_request]
+on: [push]
 #  release:
 #    types: [published]
 
