### 2023-11-08 afcfaf8
Update pyproject.toml
Attempt fixing build/deploy TestPyPI
