+++ pyproject.toml
@@ -52,4 +52,4 @@ libssa-gui = "libssa.libssa2:spawn_gui"
 # I highly recommend using this!
 [tool.setuptools_scm]
 version_scheme = "no-guess-dev"
-local_scheme = "node-and-date"
+local_scheme = "no-local-version"  # "node-and-date"
