### 2023-11-08 kstenio/libssa#2 @kstenio
Sorry for taking so long for the reply. Either way, after think more on this subject, I realized it is indeed a valid point. Also, lately a lot of people was installing LIBSsa because a LIBS class happening now, and a LOT of people had problems... I just thought "it would be actually helpful just using `pip install libssa`".

So, I did accept the PR, and did some minor tweaks to upload to TestPyPI, but got an error. It is related to PEP 440 constraint (version was auto set to `libssa-0.0.post1.dev1+g3ae7e78` during [GitHub actions](https://github.com/kstenio/libssa/actions/runs/6792255766/job/18465225272)). Any ideas on what could be the source? I think it is something related to setuptools-scm, but not sure.
