+++ pyproject.toml
@@ -23,7 +23,7 @@ dependencies = [
     "scipy~=1.11.1",
     "pandas~=2.0.3",
     "scikit-learn~=1.3.0",
-    "PySide6~=6.4.3",
+    "PySide6<=6.4.3",
     "pyqtgraph~=0.13.3",
     "openpyxl~=3.1.2",
     "psutil~=5.9.5",
