### 2023-11-08 faa7904
Update pyproject.toml
