+++ pyproject.toml
@@ -20,7 +20,7 @@ classifiers = [
 ]
 dependencies = [
     "numpy~=1.25.1",
-    "scipy~=1.11.1",
+    "scipy<=1.11.1",
     "pandas~=2.0.3",
     "scikit-learn~=1.3.0",
     "PySide6<=6.4.3",
