### 2023-11-08 kstenio/libssa#2 @kstenio
When trying to install, using `pip install --index-url https://test.pypi.org/simple/ libssa` I've found some issues:

1) Conflict using `PySide6~=6.4.3`. Changed to `<=`
2) Conflict regarding version. Changed  "no-guess-dev" to `version_scheme = "guess-next-dev"`
3) After this, pip entered in some sort of conflict regarding the two versions `ERROR: Cannot install libssa==0.0.post1.dev1 and libssa==0.1.dev1 because these package versions have conflicting dependencies.`. This is really odd to me... I mean, why is pip trying to install two versions of the same package? :thinking:
4) Then, I tried `pip install --index-url https://test.pypi.org/simple/ "libssa==0.1.dev1"` and got another error regarding version for scipy (scipy~=1.11.1). Updated to `<=` also
5) After another commit, another error in [actions regarding versions](https://github.com/kstenio/libssa/actions/runs/6792650058/job/18466274298): `HTTPError: 400 Bad Request from https://test.pypi.org/legacy/ File already exists`. Version was the previous one, `0.1.dev1`. Meaning I'm still have to get used with setuptools-scm...
