### 2023-11-08 kstenio/libssa#2 @domna
> Sorry for taking so long for the reply. Either way, after think more on this subject, I realized it is indeed a valid point. Also, lately a lot of people was installing LIBSsa because a LIBS class happening now, and a LOT of people had problems... I just thought "it would be actually helpful just using `pip install libssa`".
>
> So, I did accept the PR, and did some minor tweaks to upload to TestPyPI, but got an error. It is related to PEP 440 constraint (version was auto set to `libssa-0.0.post1.dev1+g3ae7e78` during [GitHub actions](https://github.com/kstenio/libssa/actions/runs/6792255766/job/18465225272)). Any ideas on what could be the source? I think it is something related to setuptools-scm, but not sure.

Yes, the problem is that you removed that the action is only run on release. With a release you create a git tag on which the version is based. setuptools-scm uses this tag to build the version from, e.g., if you create a tag `v0.0.1` setuptools-scm would create the version libssa-0.0.1 for you. The `post1.dev1+g3ae7e78` part is only added because setuptools-scm thinks that this is a development version (when installed directly from the repo for example) and tries to reflect the version state. The version `0.0` is because setuptools-scm is not able to find *any* tag representing a version (don't know why it didn't take your v2.1.0 tag though).
>
> [Edit]
>
> Just a few moments after posting, I found out that editing `local_scheme = "no-local-version"` instead of "node-and-date" was enough. It is now on [TestPyPI](https://test.pypi.org/project/libssa/)!

Yes this works but just removes the `post1....` part of the version (this is called the local_scheme). It still is **not** correct versioning.

You should re-add that the version is only created on release and only release with an appropriate git tag (you have to create a tag when releasing). No worries you can also remove releases if you only want to create them for testing.
