### 2023-11-08 kstenio/libssa#2 @domna
Generally, I think there is something going wrong due to the problems you had in packaging and versioning. You first should try to install your package locally (i.e., git cloning the repo and then do a `pip install .` or [`pip install -e .`](https://setuptools.pypa.io/en/latest/userguide/development_mode.html#development-mode-a-k-a-editable-installs))
For me it works all fine locally with the provided versions.

> When trying to install on **Python 3.9.18**, using `pip install --index-url https://test.pypi.org/simple/ libssa` I've found some issues:
>
> 1. Conflict using `PySide6~=6.4.3`. Changed to `<=`

I think it's related to the malfunctioning `libssa` versioning.
> 2. Conflict regarding version. Changed  "no-guess-dev" to `version_scheme = "guess-next-dev"`

I think this is related to setuptools-scm's inability to properly extract a version. I think which you choose should be a decision for you and your repository not imposed by if it works or not. This should also just apply to development installs with `no-guess-dev` not inferring the next version (i.e., latest tag `v0.0.1`, dev version something like `0.0.1.post1.dev...`) while `guess-next-dev` does infer (latest tag `v0.0.1`, dev version `0.0.2.post1.dev....`).
> 3. After this, pip entered in some sort of conflict regarding the two versions `ERROR: Cannot install libssa==0.0.post1.dev1 and libssa==0.1.dev1 because these package versions have conflicting dependencies.`. This is really odd to me... I mean, why is pip trying to install two versions of the same package? 🤔

This is odd and I think it's again related to the wrong package versioning. Normally, pip would remove the old package and install the new one. I think you just had installed it from previous testing but pip had some problems due to the wrong version formatting (I think pip would keep the newer version so it needs to figure out which one is newer).
> 4. Then, I tried `pip install --index-url https://test.pypi.org/simple/ "libssa==0.1.dev1"` and got another error regarding version for scipy (scipy~=1.11.1). Updated to `<=` also
> 5. After another commit, another error in [actions regarding versions](https://github.com/kstenio/libssa/actions/runs/6792650058/job/18466274298): `HTTPError: 400 Bad Request from https://test.pypi.org/legacy/ File already exists`. Version was the previous one, `0.1.dev1`. Meaning I'm still have to understand/get used with setuptools-scm...

Yes, pypa is not designed to replace the version which already exists. Once you release it there the version tag cannot be replaced (you can revoke it though - but best practice is to create a newer version in case there was anything wrong with this version, except for security issues then it's good to remove the problematic version from pypi) to guarantee that users always get the supplied version. It's again a problem that your version does not change, with a release your version would change for every time you push something to pypa (testing).

You can consult `setuptools-scm`'s [documentation(https://setuptools-scm.readthedocs.io) for more information. However, their documentation has not been the best in the past (especially for the basic ideas of setuptools-scm). Currently, they are updating the docs. It's still in the process so it might be confusing or incomplete.
