### 2023-11-08 214a806
Addressing issue #4 (packaging)
publish-as-package.yaml
* Rollback to previous setting (by domna)

libssa2.py
* Indentation fix
* Replaced README.md to README_GUI.md (added new) into libssa folder (module)

MANIFEST.in
* Added markdown files as include

pyproject.toml
* Replaced ~= to <= for dependencies
* Added entry for packages['libssa'] in tool.setuptools
* Rollback to version_scheme = "no-guess-dev" and local_scheme = "node-and-date"
