+++ libssa/libssa2.py
@@ -18,1629 +18,1628 @@
 
 
 # Imports
-import sys
-
 try:
-    import lzma
-    import pickle
-    import libssa.env.export as export
-    from time import time
-    from os import listdir
-    from pathlib import Path
-    from pandas import DataFrame
-    from markdown import markdown
-    from datetime import datetime
-    from traceback import print_exc
-    from psutil import virtual_memory
-    from libssa.env.spectra import Spectra, Worker
-    from libssa.env.gui.libssagui import LIBSsaGUI, changestatus
-    from libssa.env.imports import load, outliers, refcorrel, domulticorrel
-    from libssa.env.functions import (
-        isopeaks,
-        fitpeaks,
-        linear_model,
-        zeros,
-        column_stack,
-        pca_do,
-        pca_scan,
-        pls_do,
-        tne_do,
-        array,
-    )
-    from PySide6.QtWidgets import (
-        QApplication,
-        QMessageBox,
-        QMainWindow,
-        QTableWidgetItem,
-    )
-    from PySide6.QtCore import QThreadPool, QObject, QCoreApplication, Qt
+	import sys
+	import lzma
+	import pickle
+	import libssa.env.export as export
+	from time import time
+	from os import listdir
+	from pathlib import Path
+	from pandas import DataFrame
+	from markdown import markdown
+	from datetime import datetime
+	from traceback import print_exc
+	from psutil import virtual_memory
+	from libssa.env.spectra import Spectra, Worker
+	from libssa.env.gui.libssagui import LIBSsaGUI, changestatus
+	from libssa.env.imports import load, outliers, refcorrel, domulticorrel
+	from libssa.env.functions import (
+		isopeaks,
+		fitpeaks,
+		linear_model,
+		zeros,
+		column_stack,
+		pca_do,
+		pca_scan,
+		pls_do,
+		tne_do,
+		array,
+	)
+	from PySide6.QtWidgets import (
+		QApplication,
+		QMessageBox,
+		QMainWindow,
+		QTableWidgetItem,
+	)
+	from PySide6.QtCore import QThreadPool, QObject, QCoreApplication, Qt
 except (ImportError, ImportWarning) as err:
-    err_msg = str(err)
-    print(
-        f"\nYou have missing libraries to install.\n\n"
-        f"\tError message: {err_msg}\n\n"
-        f"Check the README.md for extra info.\n"
-    )
-    if "opengl" in err_msg.lower():
-        print(
-            'If you are under Linux Mint 20+ (or Ubuntu 20.04+), try running: "apt install libopengl0"'
-        )
-    elif "xcb" in err_msg.lower():
-        print(
-            'If you are under Linux Mint 20+ (or Ubuntu 20.04+), try running: "apt install libxcb-xinerama0 libxcb-cursor0"'
-        )
-    print_exc()
-    sys.exit(1)
+	err_msg = str(err)
+	print(
+		f"\nYou have missing libraries to install.\n\n"
+		f"\tError message: {err_msg}\n\n"
+		f"Check the README.md for extra info.\n"
+	)
+	if "opengl" in err_msg.lower():
+		print(
+			'If you are under Linux Mint 20+ (or Ubuntu 20.04+), try running: "apt install libopengl0"'
+		)
+	elif "xcb" in err_msg.lower():
+		print(
+			'If you are under Linux Mint 20+ (or Ubuntu 20.04+), try running: "apt install libxcb-xinerama0 libxcb-cursor0"'
+		)
+	print_exc()
+	sys.exit(1)
 
 
 # LIBSsa main class
 class LIBSSA2(QMainWindow):
-    """
-    This is LIBSsa main APP.
+	"""
+	This is LIBSsa main APP.
 
[…]
