### 2023-11-09 kstenio/libssa#2 @kstenio
Hi! I believe now the packaging in almost complete. In my last [commit](https://github.com/kstenio/libssa/commit/214a8060c52a56499d54f0eec8f805fa4d9bc84c) I managed to address most of the problems. Using `pip install .` was crucial for testing purposes, and I had to add 1) `recursive-include libssa *.md` in **MANIFEST.in** file and also 2) `packages = ["libssa"]` in **pyproject.toml**.

The include 2) was tricky, because it appeared 1st as an install error (`error: Multiple top-level packages discovered in a flat-layout: ['lvenv', 'libssa'].`). Then solved by adding `py-modules = []`, just to a 2nd error appear as an import error when loading **libssa-gui** script (`from libssa.libssa2 import spawn_gui | ModuleNotFoundError: No module named 'libssa'`). Besides, I also rolled back previous changes (including the action to deploy only on publish).

Then, finally was able to install from testpypi. Could you run in a fresh environment `pip install --index-url https://test.pypi.org/simple/ libssa` and check if works?

I noticed SVG links are broken now. I'll have to think on how properly implement it, since now this is done inside the UI file (PySide6-designer). One option would be using setIcon for GUI objects. I also have to include the changes from **joss** branch into **master** (all changes I did regarding #3 was in **joss**, not on _master_).
