### 2023-11-09 kstenio/libssa#3 @kstenio
> ... gui-script gives users a simple executable script from which they can start the program.

I have to admit: the GUI script works like a charm!

> I think this is also a point for packaging, as you can include the supported python versions in the package metadata

I believe this is happening now, right? In pyproject.toml?  I just did not get how this badge works. I have to reference the pypi repository in README.md?

Lastly, the changes in [README.md](https://github.com/kstenio/libssa/blob/joss/README.md) are sufficient for this issue? (changes were made in the **joss** branch, not _master_). After your feedback in #5 I'll add also the revised install instructions into the readme file (regarding installing from pip/pypi).
