### 2023-11-09 kstenio/libssa#2 @domna
I still get an error when I try to install it from `test.pypi`:
```
ERROR: Cannot install libssa==0.0.post1.dev1, libssa==0.1.dev1 and libssa==2.1.0.post0.dev0 because these package versions have conflicting dependencies.

The conflict is caused by:
    libssa 2.1.0.post0.dev0 depends on pyqtgraph<=0.13.3
    libssa 0.1.dev1 depends on psutil~=5.9.5
    libssa 0.0.post1.dev1 depends on psutil~=5.9.5

To fix this you could try to:
1. loosen the range of package versions you've specified
2. remove package versions to allow pip attempt to solve the dependency conflict

ERROR: ResolutionImpossible: for help visit https://pip.pypa.io/en/latest/topics/dependency-resolution/#dealing-with-dependency-conflicts
```

I'm not sure if you're specifying the package dependencies correctly and I think it should actually detect the package automatically. See [here](https://setuptools.pypa.io/en/latest/userguide/quickstart.html#package-discovery) for documentation on package discovery. It does have an include/exclute syntax and for your package you could do:
```
[tool.setuptools.packages.find]
include = ["libssa"]
```
then it should only include the `libssa` directory.
I think that changing this could resolve your problems.

> I noticed SVG links are broken now. I'll have to think on how properly implement it, since now this is done inside the UI file (PySide6-designer).

Yes this was one thing I also found but left if for you to do the fixes as you are probably much faster at finding the appropriate code locations to change.

> One option would be using setIcon for GUI objects. I also have to include the changes from joss branch into master (all changes I did regarding https://github.com/kstenio/libssa/issues/3 was in joss, not on master).

Yes, I think all your updates should go directly to master. The joss branch should only just host the paper.
