### 2023-11-09 kstenio/libssa#3 @domna
> > ... gui-script gives users a simple executable script from which they can start the program.
>
> I have to admit: the GUI script works like a charm!

Nice, I'm glad it works nicely for you.
>
> > I think this is also a point for packaging, as you can include the supported python versions in the package metadata
>
> I believe this is happening now, right? In pyproject.toml? I just did not get how this badge works. I have to reference the pypi repository in README.md?
>

Yes, you have to include something like this:

which will resolve the badge for you. However, it only will work as soon as the package is published on pypi as this loads the metadata from there.

> Lastly, the changes in [README.md](https://github.com/kstenio/libssa/blob/joss/README.md) are sufficient for this issue? (changes were made in the **joss** branch, not _master_). After your feedback in #5 I'll add also the revised install instructions into the readme file (regarding installing from pip/pypi).

Yes, thank you. I read through the README once more and found some minor issues:

- Do you want to keep the version on top of the README file? With your github ci you would be able to release directly from github and you would always need to update the version there manually (don't know if there is a github variable to get this into the readme automatically). I always keep forgetting updating such things and therefore I don't include such things hard coded anywhere. So that's what I would recommend, but if you'd like to keep it it's fine for me.
- ```remove of outliers using SAM or MAD algorithms``` should be ```removal of ...` or `remove outliers`
- I'm still not particular fond of listing and linking all the dependencies as it's just crowding the readme for me. But if you like to keep them this is no blocker for the issue.
- I would omit the user here: `pip install -r requirements.txt --user`.
- The link to your thesis shows `403 Forbidden` for me. I know it worked previously. Could you give the link a check?
- As soon as the packaging is ready I would give the README an update once more, describing adding a `pip install libssa` for people just wanting to use the package and making your current install process a development install instruction.
