### 2023-11-09 kstenio/libssa#2 @kstenio
This is really odd... I tried retiring old packages, but looks live pip is trying to install 2 versions of libssa at once :thinking:

What I did locally and worked nicely was:

```bash
git clone https://github.com/kstenio/libssa
cd libssa
python3.9 -m venv libssa-venv
source libssa-venv/bin/activate
pip cache purge
pip install .
libssa-gui
```

I think the issue is related with package naming convention. I'll try to up a release **v2.1.1**, and then a release **v2.1.2** later, test, and see if works
