### 2023-11-10 c4cb36d
Addressing issue #4 (packaging-2)
pyproject.toml
* Added description
* Removed license field (as it is in classifiers)
* Added setuptools find entry
