### 2023-11-10 ab6e761
Addressing issue #4 (packaging-2)
pyproject.toml
* Added description
* Removed license field (as it is in classifiers)
* Added setuptools find entry
