### 2023-11-10 7f68367
Addressing issue #4 (packaging-3)
publish-as-package.yaml
* Added PyPI publish
