+++ .github/workflows/publish-as-package.yaml
@@ -8,7 +8,7 @@
 
 name: Upload Python Package [Test]
 
-on: # [push]
+on:
   release:
     types: [published]
 
@@ -34,6 +34,9 @@ jobs:
     - name: Publish package to TestPyPI
       uses: pypa/gh-action-pypi-publish@release/v1
       with:
-        user: __token__
         password: ${{ secrets.TEST_PYPI_API_TOKEN }}
         repository-url: https://test.pypi.org/legacy/
+    - name: Publish package to PyPI
+      uses: pypa/gh-action-pypi-publish@release/v1
+      with:
+        password: ${{ secrets.PYPI_API_TOKEN }}
