### 2023-11-10 kstenio/libssa#2 @kstenio
After removing some releases and tags, and then creating the releases [v2.1.01](https://github.com/kstenio/libssa/releases/tag/v2.1.01) and [v2.1.02](https://github.com/kstenio/libssa/releases/tag/v2.1.02), I was still getting the errors... It was regarding _pandas_ this time:

```bash
pip install --index-url https://test.pypi.org/simple/ libssa
Looking in indexes: https://test.pypi.org/simple/
Collecting libssa
  Downloading https://test-files.pythonhosted.org/packages/8e/71/bd11882adb43ff52b805167db1020932317d93ff7d73d80beb4b5f5d0cd0/libssa-2.1.1-py3-none-any.whl (205 kB)
     ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 205.2/205.2 kB 1.2 MB/s eta 0:00:00
ERROR: Could not find a version that satisfies the requirement pandas<=2.0.3 (from libssa) (from versions: none)
ERROR: No matching distribution found for pandas<=2.0.3
```

Which I found really odd (**again!**). Also, installing manually was working, with `pip install .`, so I thought: maybe pandas was **NOT** in the TestPyPI! And, as a matter of fact, it wasn't! So, I've decided adding a secret for PyPI, included a [publish to PyPI in a new commit](https://github.com/kstenio/libssa/commit/7f68367361982bbfcf7d3f2c6ffba2228a7cfed6), did a new [v2.1.03](https://github.com/kstenio/libssa/releases/tag/v2.1.03) release, used `pip install libssa` in a fresh sourced venv and voila! Working!
