### 2023-11-10 kstenio/libssa#2 @domna
Yes, I think the problem was your referral of the index url. You specified pip to *only* look at test.pypi, but there is no pandas available on test.pypi. The command `pip install --extra-index-url https://test.pypi.org/simple libssa` works, too. This just adds test.pypi as extra url but still uses normal pypi for package resolution.

Of course publishing directly on pypi works as well. Great that it works now 🎉
