### 2023-11-11 86898dc
Addressing issues #3 and #4 (packaging-4)
README.md and README_GUI.md
* Improved files based on text on _joss_ branch

CONTRIBUTING.md and FEATURES.md
* Added contributing guidelines and features in need
* Moved features text to another branch

libssa2.py
* Removed getting version from html file
* Added mode to insert pictures in about

./libssa/pic/examples
* Added pictures to improve README.md and FEATURES.md
