+++ .github/workflows/publish-as-package.yaml
@@ -6,7 +6,7 @@
 # separate terms of service, privacy policy, and support
 # documentation.
 
-name: Upload Python Package [Test]
+name: Upload Python Package [TestPyPI and PyPI]
 
 on:
   release:

[…]
