### 2023-11-11 7ec0ac3
Addressing issues #3 and #4 (packaging-5)
libssa.svg
* Added small white shadow

README.md
* Added raw path for pictures
