+++ README.md
@@ -1,11 +1,12 @@
 <p align="center">
-	<img alt="libssa_logo" src="./libssa/pic/libssa.svg" width="500em">
+	<img alt="libssa_logo" src="https://raw.githubusercontent.com/kstenio/libssa/master/libssa/pic/libssa.svg" width="500em">
 </p>
 
 # LIBSsa2 - Laser Induced Breakdown Spectroscopy spectra analyzer
 
 ![PyPI - Python Version](https://img.shields.io/pypi/pyversions/libssa?color=5500d4)
 ![PyPI - Version](https://img.shields.io/pypi/v/libssa?color=6c037f)
+![PyPI - License](https://img.shields.io/pypi/l/libssa?color=902da6)
 
 **LIBSsa 2** is free software, licensed under the **GNU Affero General Public License version 3**, 
 and written in **Python 3.x** for analyzing **LIBS spectra**. It can read several formats of data inside raw/text files,
@@ -13,11 +14,11 @@ import the spectra, and perform multiple analyses on the data. The environment (
 into an encapsulated **lb2e** file.
 
 <p align="center">
-	<img alt="libssa_mainwindow" src="libssa/pic/examples/libssa.png" width="900em"><br>
+	<img alt="libssa_mainwindow" src="https://github.com/kstenio/libssa/raw/master/libssa/pic/examples/libssa.png" width="900em"><br>
 	<i>LIBSsa main window</i>
 </p>
 
-For a full list of features implemented in LIBSsa, you may check the [FEATURES.md](./FEATURES.md) file but, in short, you can: 
+For a full list of features implemented in LIBSsa, you may check the [FEATURES.md](https://raw.githubusercontent.com/kstenio/libssa/master/FEATURES.md) file but, in short, you can: 
 remove of outliers using SAM or MAD algorithms, perform full spectrum normalization (FSN), create correlation spectrum, 
 do peak isolation, peak fitting (Gaussian, Lorentzian, and Voigt), univariate linear models, multivariate partial least 
 squares regression (PLSR), principal components analysis (PCA), and obtain plasma temperature and electron density values 
@@ -25,7 +26,7 @@ using Saha-Boltzmann (modified to LIBS) equation.
 
 Bellow are presented some of the graphics generated by LIBSsa.
 <p align="center">
-	<img alt="libssa_graphics" src="libssa/pic/examples/montage.png" width="900em"><br>
+	<img alt="libssa_graphics" src="https://github.com/kstenio/libssa/raw/master/libssa/pic/examples/montage.png" width="900em"><br>
 	<i>Some treatments available in LIBSsa: peak isolation (a), peak fitting (b), PCA (c), and Saha-Boltzmann plot (d)</i>
 </p>
 
@@ -80,7 +81,7 @@ Lastly, advanced users are encouraged to use [venv](https://docs.python.org/3/li
 ## Contributing
 
 If you wish to contribute to the development of LIBSsa, please check for the guidelines in the 
-[CONTRIBUTING.md](./CONTRIBUTING.md) file.
+[CONTRIBUTING.md](https://raw.githubusercontent.com/kstenio/libssa/master/CONTRIBUTING.md) file.
 
 ## Warranty
 
