+++ libssa/README_GUI.md
@@ -1,12 +1,8 @@
 <p align="center">
-	<img alt="libssa_logo" src="./pic/libssa.svg" width="500em">
+	<img alt="libssa_logo" src="https://raw.githubusercontent.com/kstenio/libssa/master/libssa/pic/libssa.svg" width="500em">
 </p>
 
 # LIBSsa2 - Laser Induced Breakdown Spectroscopy spectra analyzer
-
-![PyPI - Python Version](https://img.shields.io/pypi/pyversions/libssa?color=5500d4)
-![PyPI - Version](https://img.shields.io/pypi/v/libssa?color=6c037f)
-![PyPI - License](https://img.shields.io/pypi/l/libssa?color=902da6)
 
 **LIBSsa 2** is free software, licensed under the **GNU Affero General Public License version 3**, 
 and written in **Python 3.x** for analyzing **LIBS spectra**. It can read several formats of data inside raw/text files,
@@ -14,11 +10,11 @@ import the spectra, and perform multiple analyses on the data. The environment (
 into an encapsulated **lb2e** file.
 
 <p align="center">
-	<img alt="libssa_mainwindow" src="pic/examples/libssa.png" width="900em"><br>
+	<img alt="libssa_mainwindow" src="https://github.com/kstenio/libssa/raw/master/libssa/pic/examples/libssa.png" width="900em"><br>
 	<i>LIBSsa main window</i>
 </p>
 
-For a full list of features implemented in LIBSsa, you may check the [FEATURES.md](./FEATURES.md) file but, in short, you can: 
+For a full list of features implemented in LIBSsa, you may check the [FEATURES.md](https://raw.githubusercontent.com/kstenio/libssa/master/FEATURES.md) file but, in short, you can: 
 remove of outliers using SAM or MAD algorithms, perform full spectrum normalization (FSN), create correlation spectrum, 
 do peak isolation, peak fitting (Gaussian, Lorentzian, and Voigt), univariate linear models, multivariate partial least 
 squares regression (PLSR), principal components analysis (PCA), and obtain plasma temperature and electron density values 
@@ -26,7 +22,7 @@ using Saha-Boltzmann (modified to LIBS) equation.
 
 Bellow are presented some of the graphics generated by LIBSsa.
 <p align="center">
-	<img alt="libssa_graphics" src="pic/examples/montage.png" width="900em"><br>
+	<img alt="libssa_graphics" src="https://github.com/kstenio/libssa/raw/master/libssa/pic/examples/montage.png" width="900em"><br>
 	<i>Some treatments available in LIBSsa: peak isolation (a), peak fitting (b), PCA (c), and Saha-Boltzmann plot (d)</i>
 </p>
 
@@ -34,7 +30,7 @@ Finally, users can also export the results of the analysis in many common format
 
 ## Setting up the environment
 
-You will need to install Python 3 to use the program (tested on 3.9 with current libraries). Python can be obtained from 
+You will need to install Python 3 to use the program (tested on 3.9 and 3.10 with current libraries). Python can be obtained from 
 the official website, [https://www.python.org/](https://www.python.org/) and, if you have any additional inquiries, there are 
 official instructions on how install Python on [Linux Distributions](https://docs.python.org/3/using/unix.html), 
 [Windows](https://docs.python.org/3/using/windows.html) and [macOS](https://docs.python.org/3/using/mac.html).
@@ -81,7 +77,7 @@ Lastly, advanced users are encouraged to use [venv](https://docs.python.org/3/li
 ## Contributing
 
 If you wish to contribute to the development of LIBSsa, please check for the guidelines in the 
[…]
