### 2023-11-11 b35a584
Addressing issues #3 and #4 (packaging-5)
libssa.svg
* Another small adjustment with the white shadow

README.md
* Contributing and Features now references blob instead raw content (on GitHub)
