### 2023-11-11 2c29542
Addressing issues #3 and #4 (packaging-6)
libssa.svg
* Rollback...

README.md
* Updated/included instructions for PyPI install
