### 2023-11-11 kstenio/libssa#2 @kstenio
> Yes, I think the problem was your referral of the index url. You specified pip to _only_ look at test.pypi, but there is no pandas available on test.pypi. The command `pip install --extra-index-url https://test.pypi.org/simple libssa` works, too. This just adds test.pypi as extra url but still uses normal pypi for package resolution.

I see... so the problem was the `-i` option

> Of course publishing directly on pypi works as well. Great that it works now 🎉

Today I did some cleanup/testing/fine tune in the repository. Tested until release [v2.1.07](https://github.com/kstenio/libssa/releases/tag/v2.1.07) and all seems fine with packaging (still, I'll clean up dev tags/releases after this issue is closed). [TestPyPI](https://test.pypi.org/project/libssa/) page also have the new info from the README. I believe it is all set up now. If this is also your opinion, please feel free to close the issue :smile:
