### 2023-11-11 kstenio/libssa#3 @kstenio
> Yes, you have to include something like this:
>
> ```
> ```

>     * Do you want to keep the version on top of the README file? With your github ci you would be able to release directly from github and you would always need to update the version there manually (don't know if there is a github variable to get this into the readme automatically). I always keep forgetting updating such things and therefore I don't include such things hard coded anywhere. So that's what I would recommend, but if you'd like to keep it it's fine for me.

I removed. It is really better now with the shield metadata

>     * `remove of outliers using SAM or MAD algorithms` should be ```removal of ...`or`remove outliers`

Thanks for pointing this out. It was in fact an error. I removed the "of".

>     * I'm still not particular fond of listing and linking all the dependencies as it's just crowding the readme for me. But if you like to keep them this is no blocker for the issue.

I included two options for installing libssa: 1) using pip and 2) using git clone. I liked the way it is now, as I feel like talking to both audiences: end users and entry-level developers.

>     * I would omit the user here: `pip install -r requirements.txt --user`.

Yes. I did experienced an error regarding this option recently. Removed now.

>     * The link to your thesis shows `403 Forbidden` for me. I know it worked previously. Could you give the link a check?

This is odd. Maybe was a problem with the university server. I've checked now and seems fine.

>     * As soon as the packaging is ready I would give the README an update once more, describing adding a `pip install libssa` for people just wanting to use the package and making your current install process a development install instruction.

Yes, this info in now in the [Simple Install](https://github.com/kstenio/libssa#1-simple-install-pypi-version) section :smiley:

Also, in the last commits I did a lot regarding the merge of changes only previously in **joss** branch. Pictures links uses raw links instead of local ones (PyPI was not showing images correctly). Now, the readme looks way better than before!
