### 2023-11-12 kstenio/libssa#2 @domna
Yes it looks all nice. You may consider adding you development dependencies (pytest etc.) to an extra installs section in pyproject.toml which would help the install project. But I think the main part of packaging is done so I'm closing this.
