### 2023-11-12 28cb918
Addressing issues #3 (readme-1)
README.md
* Added extra badges
* Changed to shields.io do badgen.net
* Updated/included instructions for PIP to install LIBSsa in editable mode
