### 2023-11-12 kstenio/libssa#3 @kstenio
> Yes, this looks very nice now. It even fits your programs color scheme :)

You noticed! I've also included some other tags and tried using [badgen.net](https://badgen.net/) since [shields.io](https://shields.io) went down for some unknown problem.

> This installs the package in editable mode and the procedure for starting the program is the same as with the packaged install. In editable mode every change to the source code is also reflected in the package.

This was a great tip! I liked, and since the install is supported by pip, the auto check for Python version is automatic.

>  It would also install the dependencies for you so you don't have to do the manual requirements install

Yes, I thought so. The updated readme does not references `pip install -r requirements.txt` anymore. Do you recommend removing the file itself from the repository root? (usualy, IDEs would use the file for venv automation)

> Edit: Another small point I found: I find it always nice when the package badge with the version directly link to pypi, so when you click it you directly land on your projects page on pypi.

I did not figure out how to do this using markdown. I think I could use HTML <img> tag though.
