### 2023-11-13 d53f9d9
Addressing issue #3 (readme-2)
publish-as-package.yaml
* Re-enabled publish to PyPI

pyproject.toml
* Added dependencies for dev install: `pip install libssa[dev]`

README.md
* Added links for badges
