### 2023-11-13 kstenio/libssa#3 @kstenio
> I like to keep them. I do however use it to generate a fixed environment with all the dependencies (and sub-dependencies) pinned to a fixed version. That way you can use it to re-generate a fixed environment. I normally use pip-compile (from the package [pip-tools](https://github.com/jazzband/pip-tools)) for this.

Ok. I've decided to keep the `requirements.txt` file as well.

> It should work with markdown as well. Try doing something like this:

Thanks for the tip! I've included the PyPI and other links for the project now!
