### 2023-11-16 1b55489
Addressing issue #3 (readme-3)
libssagui.ui and libssagui.py
* Added new element in menu for loading sample data

libssa2.py
* Now imports tarfile and tempfile libs
* New tempfolder class variable created to save extracted sample spectra
* New connection from menu to loadsample_spectra
* Method loadsample_spectra now extracts sampled synthetic_samples_ultra-low-res-spectrometer_c-model.tar.gz file to self.tempfolder, then triggers spopen and spload
* Method spopen now works in empty path mode (normal) and non empty (sample data)
* Spload does self.tempfolder cleanup
