### 2023-11-16 kstenio/libssa#3 @kstenio
Hello @domna. In My last [commit](https://github.com/kstenio/libssa/commit/1b554893bf2a91ad0b0179fac91338896c0203ca), I did include the sample data as previously requested. Now, from the file menu, the user may import a simple set of samples to explore LIBSsa on its own.

I believe all main concerns regarding this issue are concluded. Please, close the thread if you agree.
