### 2023-11-16 #5961 @kstenio
> 👋🏻 I just wanted to check in on the progress of the reviews here. It looks like there are active discussions going on in the issues linked above 👍🏻

Hi @csoneson, thank you very much for the continuous support. In fact I've been working continuously with @domna on the issues, and some key aspects of the application are much better now. @Tillsten also gave nice insights.
