### 2023-11-19 kstenio/libssa#3 @domna
Yes, I think this issue has been resolved. Thanks!
