### 2023-11-27 0853cb9
Addressing issue #7
libssa2.py
* Added false condition to path in spopen

README_GUI.md
* Changed pictures path to relative (tested on IDE)
