+++ libssa/README_GUI.md
@@ -1,5 +1,5 @@
 <p align="center">
-	<img alt="libssa_logo" src="https://raw.githubusercontent.com/kstenio/libssa/master/libssa/pic/libssa.svg" width="500em">
+	<img alt="libssa_logo" src="./pic/libssa.svg" width="500em">
 </p>
 
 # LIBSsa2 - Laser Induced Breakdown Spectroscopy spectra analyzer
@@ -10,7 +10,7 @@ import the spectra, and perform multiple analyses on the data. The environment (
 into an encapsulated **lb2e** file.
 
 <p align="center">
-	<img alt="libssa_mainwindow" src="https://github.com/kstenio/libssa/raw/master/libssa/pic/examples/libssa.png" width="900em"><br>
+	<img alt="libssa_mainwindow" src="./pic/examples/libssa.png" width="900em"><br>
 	<i>LIBSsa main window</i>
 </p>
 
@@ -22,7 +22,7 @@ using Saha-Boltzmann (modified to LIBS) equation.
 
 Bellow are presented some of the graphics generated by LIBSsa.
 <p align="center">
-	<img alt="libssa_graphics" src="https://github.com/kstenio/libssa/raw/master/libssa/pic/examples/montage.png" width="900em"><br>
+	<img alt="libssa_graphics" src="./pic/examples/montage.png" width="900em"><br>
 	<i>Some treatments available in LIBSsa: peak isolation (a), peak fitting (b), PCA (c), and Saha-Boltzmann plot (d)</i>
 </p>
 
