### 2023-12-04 #5961 @kstenio
Hello everyone, thanks for the continuous support!

Just a small reminder regarding **_Data sharing_**: the paper itself does not contain original data, however, for the reviewers being able to explore the app and its core functionalities, I did upload some spectra in the short-lived branch [joss_sample_data](https://github.com/kstenio/libssa/tree/joss_sample_data/data) (guidelines included), as previously accorded in the [pre-review](https://github.com/openjournals/joss-reviews/issues/5805).

Also, during the resolution of issue [#3](https://github.com/kstenio/libssa/issues/3#), I've opted to include some samples [natively on the app](https://github.com/kstenio/libssa/commit/1b554893bf2a91ad0b0179fac91338896c0203ca#diff-bd9b1630e8c91b91f45cdcf381789808a511b7aa39c3e99d9e1a427b1df0a2bf), so it is much easier for anyone to explore LIBSsa now.
