### 2023-12-05 #5961 @alberto-battistel
> Thanks @alberto-battistel! Just to make sure I understand correctly - there are still a couple of unticked boxes in your checklist; are those related to the still-open issue, or did the reply to the issue address your concerns?

I closed the issue as the reply addressed my concern.
