### 2023-12-18 kstenio/libssa#4 @domna
Hey @kstenio, what is the status with the three basic tests? Do you still want to implement these for the review?
