### 2023-12-19 kstenio/libssa#4 @kstenio
Hi @domna, thank you very much for reaching out regarding this issue. Sorry taking this long to work on this, I was focusing on the other issues first, and I recently switched jobs, which left me with little time for other projects. But, I'll try to do something regarding the tests.

I've tested the gui test you've added and seems to be passing without issues:

![image](https://github.com/kstenio/libssa/assets/9257942/d8bd5020-0b37-4eb8-b1a2-084885f984b9)

As I said before, I have little no none knowledge regarding tests, but I believe a test for the [load function](https://github.com/kstenio/libssa/blob/master/libssa/env/imports.py#L30), shouldn't be too difficult to implement. My idea is to mock some data (single and multiple modes) and pass to the function. Is this the correct approach?
