### 2023-12-19 #5961 @kstenio
Hi @csoneson and @domna, thanks for the continuous support. Sorry taking so long to work on issue [#4](https://github.com/kstenio/libssa/issues/4). Tests are a difficult subject for me, so I've focused on the other issues first.

But I have some Ideas on how to implement some minor test to the application, and I hope to be able to implement it as soon as possible.
