### 2023-12-20 6c58645
Addressing issue #4 (tests)
test_load_spectra.py
* Created a simple test script for loading spectra
* It is divided in 2 parts:
    1. Mocking spectra: creates folders and files in temp directory for test (auto deleted after test)
    2. Test load function: checks if values returned from load function (from libssa.env.imports) is as expected
