### 2023-12-20 kstenio/libssa#4 @kstenio
Hi @domna. In my [last commit](https://github.com/kstenio/libssa/commit/6c586458cadceddae0f789ca69396bc78f4a2a2a) I did a test to check if the load function works as expected. I created/mocked some spectra files and ran an assert to check if values are correct. It looks as fine:

![image](https://github.com/kstenio/libssa/assets/9257942/b2cdcee2-91e0-460d-8f2d-6af68d11f0a1)

As I said before, I'm just learning how to perform unit test, so my logic might not be perfect, or even correct, but I did my best.
