### 2023-12-20 kstenio/libssa#4 @domna
Hey @kstenio,

thank you. This looks nice already. I think it would also be useful to have a simple test walking through your program. In `pytest-qt` there is the `qtbot` fixture (you can see it in the test I contributed). This can be used to control the gui, i.e., click buttons, type things in etc. [Here](https://pytest-qt.readthedocs.io/en/latest/tutorial.html) you can find more on the topic. The benefit is that you basically do an end test of your gui ensuring that everything is doing as you wish. You're tests are more like in a unit test fashion (which is fine, too), i.e., they check a particular part of your software but you wouldn't know if anything breaks around it.

My idea here was that you can simply write an easy workflow with `qtbot` from loading of your example data, triggering some analysis and then exporting it into a file. You could then check whether the file was created and is correct according to the sample data. This would already check a lot of program features and you would easily be able to check when something breaks if you change something. If you can create such a test I think this would be a really great basic test. Do you think it's feasible for you to write such a small test with qtbot?
