### 2023-12-20 kstenio/libssa#4 @kstenio
> thank you. This looks nice already.

Nice to hear! I wasn't sure if the approach was correct :sweat_smile:

> My idea here was that you can simply write an easy workflow with qtbot from loading of your example data, triggering some analysis and then exporting it into a file. You could then check whether the file was created and is correct according to the sample data. This would already check a lot of program features and you would easily be able to check when something breaks if you change something.

I see. This seems like an interesting way to check if the workflow works as expected.

> If you can create such a test I think this would be a really great basic test. Do you think it's feasible for you to write such a small test with qtbot?

I'll have to investigate better on how to use `qbot`, but looks reasonably simple. That being said, I believe I can create such test. My idea is going from loading sample data to exporting fitting areas. This will cover a huge part of the software.
