### 2023-12-22 2007c81
Addressing issue #4 (tests) [test for saving areas]
test_export_areas.py
* Created a test script to check if program reads data, export values and confirms if obtained result is consistent
* A little odd to create. QThreadPool in operations resulted in unexpected behaviour, but the qtbot waitUntil implementation was the key

imports.py
* Updated delim standard
