### 2023-12-22 kstenio/libssa#4 @kstenio
Hi @domna! I believe the test is implemented (not the prettiest, but it works). Please, check the [latest commit](https://github.com/kstenio/libssa/commit/2007c8141ec5ba18b26eb5aba6e365a5becbbfb6).

I used `qtbot` until being able to export some data from the sample spectra. Then, the test checks if values is consistent. Now, the 3 tests are passing :smile:

![image](https://github.com/kstenio/libssa/assets/9257942/d204bfb7-d69c-4a17-b37b-142a70a8958e)
