### 2023-12-22 kstenio/libssa#4 @domna
Nice, looks good. So I'm closing this issue :)
