+++ paper.bib
@@ -2,7 +2,7 @@
   title={Laser induced breakdown spectroscopy},
   author={Miziolek, Andrzej W and Palleschi, Vincenzo and Schechter, Israel},
   year={2006},
-  publisher={Cambridge university press}
+  publisher={Cambridge University Press}
 }
 
 @phdthesis{kstenio2023,
@@ -15,7 +15,7 @@
 
 @article{ferreira2011evaluation,
   title={Evaluation of laser induced breakdown spectroscopy for multielemental determination in soils under sewage sludge application},
-  author={Ferreira, Edilene Cristina and Milori, D{\'e}bora Marcondes Bastos Pereira and Ferreira, Ednaldo Jos{\'e} and Dos Santos, Larissa Macedo and Martin-Neto, Ladislau and de Ara{\'u}jo Nogueira, Ana Rita},
+  author={Ferreira, Edilene Cristina and Milori, D{\'e}bora Marcondes Bastos Pereira and Ferreira, Ednaldo Jos{\'e} and Dos Santos, Larissa Macedo and Martin-Neto, Ladislau and {de Ara{\'u}jo Nogueira}, Ana Rita},
   journal={Talanta},
   volume={85},
   number={1},
@@ -28,7 +28,7 @@
 
 @article{villas2016laser,
   title={Laser-induced breakdown spectroscopy to determine soil texture: A fast analytical technique},
-  author={Villas-Boas, Paulino Ribeiro and Romano, Renan Arnon and de Menezes Franco, Marco Aur{\'e}lio and Ferreira, Edilene Cristina and Ferreira, Ednaldo Jos{\'e} and Crestana, Silvio and Milori, D{\'e}bora Marcondes Bastos Pereira},
+  author={Villas-Boas, Paulino Ribeiro and Romano, Renan Arnon and {de Menezes Franco}, Marco Aur{\'e}lio and Ferreira, Edilene Cristina and Ferreira, Ednaldo Jos{\'e} and Crestana, Silvio and Milori, D{\'e}bora Marcondes Bastos Pereira},
   journal={Geoderma},
   volume={263},
   pages={195--202},
@@ -39,7 +39,7 @@
 }
 
 @article{ranulfi2018nutritional,
-  title={Nutritional characterization of healthy and Aphelenchoides besseyi infected soybean leaves by laser-induced breakdown spectroscopy (LIBS)},
+  title={{Nutritional characterization of healthy and Aphelenchoides besseyi infected soybean leaves by laser-induced breakdown spectroscopy (LIBS)}},
   author={Ranulfi, Anielle C and Senesi, Giorgio S and Caetano, Jonas B and Meyer, Maur{\'\i}cio C and Magalhaes, Aida B and Villas-Boas, Paulino R and Milori, Debora MBP},
   journal={Microchemical Journal},
   volume={141},
@@ -51,8 +51,8 @@
 }
 
 @article{costa2018direct,
-  title={Direct determination of Ca, K and Mg in cassava flour samples by laser-induced breakdown spectroscopy (LIBS)},
-  author={Costa, Vinicius C{\^a}mara and de Babos, Diego Victor and de Aquino, Francisco Wendel Batista and Virg{\'\i}lio, Alex and Amorim, F{\'a}bio Alan Carqueija and Pereira-Filho, Edenir Rodrigues},
+  title={{Direct determination of Ca, K and Mg in cassava flour samples by laser-induced breakdown spectroscopy (LIBS)}},
[…]
