+++ paper.md
@@ -42,7 +42,7 @@ leaves [@ranulfi2018nutritional; @stenio2022direct], fertilizers [@marangoni2016
 food [@costa2018direct], metallic alloys [@noll2018libs] and electronic waste [@costa2018laser], to name a few.
 
 The Laser-Induced Breakdown Spectroscopy spectra analyzer (**LIBSsa**) is open-source software written in Python focused on
-analyzing LIBS spectra. It combines multiple tools used in LIBS analysis into only one application, such as outliers removal,
+analyzing LIBS spectra. It combines multiple tools used in LIBS analysis into one application, such as outliers removal,
 isolation of spectral lines, curve fitting, linear models (calibration curves), principal components analysis (PCA),
 and plasma temperature and electron density calculation.
 
@@ -52,7 +52,7 @@ and plasma temperature and electron density calculation.
 
 LIBS measurement is much simpler than other standard elementary characterization techniques, such as Flame Absorption Atomic
 Spectrometry (FAAS) or Inductively Coupled Plasma Optical Emission Spectroscopy (ICP OES), since it does not require acid digestion
-for sample preparation. Furthermore, the user can obtain a sample spectrum in seconds. Nonetheless, the analysis of its signal
+for sample preparation. Furthermore, the user can obtain a sample spectrum in seconds. Nonetheless, the analysis of the resulting signal
 may be highly complex.
 
 Much of the complexity in analyzing LIBS spectra arises mainly due to matrix effects, which hinder the ability to obtain
@@ -62,8 +62,8 @@ signal-processing strategies to achieve quantitative measurements.
 In general, those working with LIBS commonly develop tools to process the measured signal and extract satisfactory results.
 However, this practice has the effect of generating highly specialized users of the technique, not only in terms of using
 instrumentation but also in scientific data analysis tools, such as [R](https://www.r-project.org/), [MATLAB](https://www.mathworks.com/products/matlab.html),
-[OriginLab](https://www.originlab.com/), [Weka](https://www.cs.waikato.ac.nz/ml/weka/), among others. Although these software
-are powerful tools for signal analysis, they are not dedicated tools for LIBS spectra analysis.
+[OriginLab](https://www.originlab.com/), [Weka](https://www.cs.waikato.ac.nz/ml/weka/), among others. Although these platforms
+provide powerful tools for signal analysis, they are not dedicated tools for LIBS spectra analysis.
 
 ![LIBSsa home screen tab with spectra loaded. Source: self-authored.\label{fig:3}](./pic/libssa.png)
 
@@ -72,28 +72,28 @@ software creation that would automate the LIBS analysis, incorporating strategie
 [@marangoni2016phosphorus; @nicolodelli2014quantification; @de2021total; @stenio2022carbon; @castro2016twelve; @stenio2022direct].
 In this way, the LIBSsa software was conceived (\autoref{fig:3}).
 
-The main purpose of LIBSsa is to help scientists in LIBS/spectroscopy field gain speed and practicality when analyzing LIBS
+The main purpose of LIBSsa is to help scientists in the LIBS/spectroscopy field gain speed and practicality when analyzing LIBS
 spectra, allowing fast assessment of which is the best calibration strategy for the sample set.
 
 # Brief software description
 
 LIBSsa is straightforward: the user selects the input source (where LIBS spectra are located), loads them into the program, and 
-does a wide range of treatments, including: removal of outliers using SAM [@keshava2004distance] or MAD [@leys2013detecting] algorithms,
[…]
