### 2023-12-31 #5961 @kstenio
> Hi @kstenio - I have gone through your submission and overall it looks good; I have opened a [PR](https://github.com/kstenio/libssa/pull/8) with a few small suggestions for the manuscript/bibliography. Please merge if you agree, then generate a new proof here with `@editorialbot generate pdf` and check that things look good.  Once that is done, please take a look at the post below for the next steps.

Hi @csoneson, thanks for being able to review the submission so soon. I already accepted the PR, and will begin to perform/check the steps now.
