### 2023-12-31 7e642ad
LIBSsa version: 2.2 [JOSS paper final #1]
libssa.png
* Replaced main app picture

README.md
* Updated version to 2.2
