### 2023-12-31 db765b1
LIBSsa version 2.2
libssa.png
* Replaced main app picture

libssagui.py
* Updated version to 2.2
