### 2023-12-31 125ee58
LIBSsa version 2.2 (untrack)
publish-as-package.yaml
* Disabled publish (for Zenodo 1st run)
