### 2023-12-31 3f1a42c
Create CITATION.cff
