### 2023-12-31 #5961 @kstenio
Hi @csoneson, I believe all is fine now:

1. Double check authors and affiliations (including ORCIDs): **I did and all is fine for [me](https://orcid.org/0000-0002-1407-9712) and [Debora](https://orcid.org/0000-0003-1253-7174)**
2. Make a release of the software with the latest changes from the review and post the version number here: **Release `v2.2.0` [created](https://github.com/kstenio/libssa/releases/tag/v2.2.0)**
3. Archive the release on Zenodo/figshare/etc and post the DOI here: **Zenodo DOI `10.5281/zenodo.10447381` [created](https://zenodo.org/doi/10.5281/zenodo.10447381)**
4. Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper: **Checked**
5. Make sure that the license listed for the archive is the same as the software license: **Checked**
