### 2024-01-11 #5961 @Kevin-Mattheus-Moerman
@kstenio As AEiC I will now help process this work for acceptance in JOSS. I have checked the paper, this review, your repository, and the archive link. All seems in order. I will now proceed to process this for acceptance.
